Image upload feature
Able to upload images to amazon aws s3 bucket /wagger/:id
Images are named 0-4 for a total of 5 images per user.
images is a property on the user model instantiated as an empty array.
on successful upload to the bucket, image url is saved to the array
and returns the updated user data