Messaging feature
Able to send message to the server and create the message on both the
user match messages and the corresponding other user's match messages.

message includes text, user reference, and time which is generated by
the server.

Server only needs to know the user_id of the poster,
the user_id of the poster's match, and the text of the message.

POST request sent to '/users/:id/matches/:match_id/messages'